More reorganization and cleanup, add comments and basic README
