+++ Cargo.toml
@@ -44,6 +44,10 @@ incremental   = false
 codegen-units = 1
 incremental   = false
 
+[profile.release.package.cw721-sylvia-base]
+codegen-units = 1
+incremental = false
+
 [profile.release]
 rpath            = false
 lto              = true
