+++ contracts/cw721-sylvia-base/Cargo.toml
@@ -1,9 +1,14 @@
 [package]
-name = "cw721-sylvia-base"
-version = "0.1.0"
-edition = { workspace = true }
-repository = { workspace = true }
-license = { workspace = true }
+name          = "cw721-sylvia-base"
+description   = "Basic implementation of cw721 NFTs using the Sylvia framework."
+version       = { workspace = true }
+edition       = { workspace = true }
+license       = { workspace = true }
+repository    = { workspace = true }
+homepage      = { workspace = true }
+documentation = { workspace = true }
+rust-version  = { workspace = true }
+
 
 [features]
 library = []
