+++ contracts/cw721-sylvia-base/README.md
@@ -1,0 +1,8 @@
+# cw721-sylvia-base
+
+This is a basic implementation of a cw721 NFT contract. It implements
+the [CW721 spec](../../packages/cw721/README.md) and is designed to
+be deployed as is, or imported into other contracts to easily build
+cw721-compatible NFTs with custom logic. It uses the [Sylvia](https://github.com/cosmwasm/sylvia) smart contract framework for easier extension.
+
+Use this as a base to build your own custom NFT contract.
