+++ contracts/cw721-sylvia-base/src/lib.rs
@@ -1,6 +1,10 @@
-mod base;
+#![doc = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/README.md"))]
+
+pub mod base;
 pub mod contract;
 mod error;
+pub mod responses;
+pub mod state;
 
 #[cfg(test)]
 mod multitest;
