+++ contracts/cw721-sylvia-base/src/responses.rs
@@ -1,0 +1,7 @@
+use cosmwasm_schema::cw_serde;
+
+/// Shows who can mint these tokens
+#[cw_serde]
+pub struct MinterResponse {
+    pub minter: Option<String>,
+}
