+++ packages/cw721/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod cw721_interface;
+mod interface;
 mod msg;
 mod query;
 mod receiver;
@@ -6,6 +6,7 @@ mod traits;
 
 pub use cw_utils::Expiration;
 
+pub use crate::interface::cw721_interface;
 pub use crate::msg::Cw721ExecuteMsg;
 pub use crate::query::{
     AllNftInfoResponse, Approval, ApprovalResponse, ApprovalsResponse, ContractInfoResponse,
